+++ database/bank_test.go
@@ -4,8 +4,6 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	dbtypes "github.com/forbole/bdjuno/v4/database/types"
-
-	bddbtypes "github.com/forbole/bdjuno/v4/database/types"
 )
 
 func (suite *DbTestSuite) TestBigDipperDb_SaveSupply() {