+++ modules/resource/module.go
@@ -1,4 +1,4 @@
-package top_accounts
+package resource
 
 import (
 	"github.com/cosmos/cosmos-sdk/codec"
@@ -10,44 +10,31 @@ import (
 )
 
 var (
-	_ modules.Module                   = &Module{}
-	_ modules.MessageModule            = &Module{}
-	_ modules.PeriodicOperationsModule = &Module{}
+	_ modules.Module        = &Module{}
+	_ modules.MessageModule = &Module{}
 )
 
-// Module represent x/gov module
+// Module represent x/resource module
 type Module struct {
 	cdc           codec.Codec
 	db            *database.Db
 	messageParser junomessages.MessageAddressesParser
-	authModule    AuthModule
-	bankModule    BankModule
-	distrModule   DistrModule
-	stakingModule StakingModule
 }
 
 // NewModule returns a new Module instance
 func NewModule(
-	authModule AuthModule,
-	bankModule BankModule,
-	distrModule DistrModule,
-	stakingModule StakingModule,
 	messageParser junomessages.MessageAddressesParser,
 	cdc codec.Codec,
 	db *database.Db,
 ) *Module {
 	return &Module{
 		cdc:           cdc,
-		authModule:    authModule,
-		bankModule:    bankModule,
-		distrModule:   distrModule,
 		messageParser: messageParser,
-		stakingModule: stakingModule,
 		db:            db,
 	}
 }
 
 // Name implements modules.Module
 func (m *Module) Name() string {
-	return "top_accounts"
+	return "resource"
 }