+++ modules/top_accounts/expected_modules.go
@@ -1,5 +1,8 @@
 package top_accounts
 
+type AuthModule interface {
+	GetTotalNumberOfAccounts(height int64) (uint64, error)
+}
 type BankModule interface {
 	UpdateBalances(addresses []string, height int64) error
 }